# frontend/src/openapi-codegen/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clochette API",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/consumable/": {
      "get": {
        "tags": [
          "consumable"
        ],
        "summary": "Read Consumables",
        "description": "Retrieve a list of consumables.",
        "operationId": "v1_read_consumables",
        "deprecated": true,
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ],
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "All"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsumableV1"
                  },
                  "title": "Response V1 Read Consumables"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/consumable/distincts/": {
      "get": {
        "tags": [
          "consumable"
        ],
        "summary": "Read Consumables Distincts",
        "description": "Retrieve a list of distinct consumables.",
        "operationId": "v1_read_consumables_distincts",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsumableV1"
                  },
                  "type": "array",
                  "title": "Response V1 Read Consumables Distincts"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          }
        },
        "deprecated": true,
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ]
      }
    },
    "/api/v1/consumable/{consumable_id}": {
      "put": {
        "tags": [
          "consumable"
        ],
        "summary": "Update Consumable",
        "description": "Update a consumable.",
        "operationId": "v1_update_consumable",
        "deprecated": true,
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ],
        "parameters": [
          {
            "name": "consumable_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Consumable Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumableUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumableV1"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/consumable/": {
      "get": {
        "tags": [
          "consumable"
        ],
        "summary": "Read Consumables",
        "description": "Retrieve a list of consumables.\n\nQuery parameters:\n    - `all`: A boolean indicating whether to return all consumables or only non-empty ones.",
        "operationId": "read_consumables",
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ],
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "All"
            }
          },
          {
            "name": "consumable_item_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Consumable Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Consumable"
                  },
                  "title": "Response Read Consumables"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "consumable"
        ],
        "summary": "Create Consumable",
        "description": "Create a consumable.",
        "operationId": "create_consumable",
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumableCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consumable"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/consumable/{consumable_id}": {
      "get": {
        "tags": [
          "consumable"
        ],
        "summary": "Read Consumable",
        "description": "Retrieve a consumable.",
        "operationId": "read_consumable",
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ],
        "parameters": [
          {
            "name": "consumable_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Consumable Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consumable"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "consumable"
        ],
        "summary": "Update Consumable",
        "description": "Update a consumable.",
        "operationId": "update_consumable",
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ],
        "parameters": [
          {
            "name": "consumable_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Consumable Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumableUpdateModify"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consumable"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/consumable/distincts/": {
      "get": {
        "tags": [
          "consumable"
        ],
        "summary": "Read Distinct Consumables",
        "description": "Retrieve a list of distinct consumables.",
        "operationId": "read_distinct_consumables",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsumableDistinct"
                  },
                  "type": "array",
                  "title": "Response Read Distinct Consumables"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsumableDistinct": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "consumableItemId": {
            "type": "integer",
            "title": "Consumableitemid"
          },
          "solded": {
            "type": "boolean",
            "title": "Solded"
          },
          "sellPrice": {
            "type": "number",
            "title": "Sellprice"
          },
          "buyPrice": {
            "type": "number",
            "title": "Buyprice"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "readOnly": true
          },
          "icon": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IconName"
              }
            ],
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "consumableItemId",
          "solded",
          "sellPrice",
          "buyPrice",
          "quantity",
          "name",
          "icon"
        ],
        "title": "ConsumableDistinct"
      },
...
      "ConsumableUpdate": {
        "properties": {
          "unitPrice": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Unitprice"
          },
          "sellPrice": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sellprice"
          }
        },
        "type": "object",
        "title": "ConsumableUpdate"
      },
      "ConsumableUpdateModify": {
        "properties": {
          "sellPrice": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sellprice"
          }
        },
        "type": "object",
        "title": "ConsumableUpdateModify"
      },
...
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Error message."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "HTTPError"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IconName": {
        "type": "string",
        "enum": [
          "Glass",
          "Beer",
          "Food",
          "Soft",
          "Barrel",
          "Misc"
        ],
        "title": "IconName"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ConsumableV1": {
        "properties": {
          "unitPrice": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Unitprice"
          },
          "sellPrice": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Sellprice"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "fkId": {
            "type": "integer",
            "title": "Fkid"
          },
          "empty": {
            "type": "boolean",
            "title": "Empty"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "readOnly": true
          },
          "icon": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IconName"
              }
            ],
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "unitPrice",
          "sellPrice",
          "id",
          "fkId",
          "empty",
          "name",
          "icon"
        ],
        "title": "Consumable"
      },
...
      "Consumable": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "consumableItemId": {
            "type": "integer",
            "title": "Consumableitemid"
          },
          "solded": {
            "type": "boolean",
            "title": "Solded"
          },
          "sellPrice": {
            "type": "number",
            "title": "Sellprice"
          },
          "buyPrice": {
            "type": "number",
            "title": "Buyprice"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "readOnly": true
          },
          "icon": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IconName"
              }
            ],
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "consumableItemId",
          "solded",
          "sellPrice",
          "buyPrice",
          "name",
          "icon"
        ],
        "title": "Consumable"
      },
      "ConsumableCreate": {
        "properties": {
          "consumableItemId": {
            "type": "integer",
            "title": "Consumableitemid"
          },
          "sellPrice": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Sellprice"
          },
          "buyPrice": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Buyprice"
          },
          "transactionId": {
            "type": "integer",
            "title": "Transactionid",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "consumableItemId",
          "sellPrice",
          "buyPrice"
        ],
        "title": "ConsumableCreate"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {
              "staff": "Manage stock and create transactions",
              "treasurer": "Manage treasury, i.e. CRUD over transactions",
              "president": "Manage users, i.e. CRUD over users"
            },
            "tokenUrl": "/api/v1/auth/login/"
          }
        }
      }
    }
  }
}